Added AI Behavior including EQS queries, Powerup mechanic, Hit Flash material damage feedback, Added a Game Mode, adeed a target dummy for demonstrating Damage Popup mechanic, added camera shake on projectile hit,
Added an SIACharacter and its child BP_MinionRanged.
Added SIAController from which we are running BT_MinionRanged.
Added BB_MinionRanged.
In BT_MinionRanged, we defined the basic AI behavior. The AI uses Query_FindNearbyLocation EQS to choose the location to which to go.
Query_FindNearbyLocation uses QueryContext_TargetActor to spawn a donut of points around the TargetActor, and it chooses the best of 25% of the points which are at least 500 cm away from the querier, preferring the shorter distance under the condition that the point has a clear line of sight to the Target Actor. The QueryContext_TargetActor provides a single actor either by taking it from the Blackboard value "TargetActor" and casting it to the Actor or, if the cast fails, it sets it to the PlayerPawn.
SIACharacter has a pawn sensing component added and has a function OnPawnSeen() which is bound to the OnSeePawn event. The OnPawnSeen() sets the Blackboard Value "TargetActor" to the seen pawn and prints a debug message.
Query_FindSpotToHeal is added and used by the BT_MinionRanged to move the AI to the nearest spot which breaks the line of sight in relation to the Player-TargetActor.
USBTService_CheckAttackRange added, which checks if the distance between the TargetActor and SIACharacter is less than the specified distance. If it is, the BlackBoard value "WithinAttackRange" is set to true, and the decorator "WithinAttackRange?" allows for the attack loop by checking "WithinAttackRange."
Added an SBTTask_RangedAttack, which is used in the attack loop in the BT_MinionRanged. It spawns the Projectile TSubclassOf<AActor> ProjectileClass and is spawned at the location of the gun socket of the Minion_Lane_Ranged_Dawn in the direction of the player with a small random Yaw offset and a small positive-only pitch offset.
SetDefaultFocus service is used in BT_MinionRanged, which rotates the AI towards the Player.
Added a USBTTask_SelfHeal, which is executed in the BT, once the FindSpotToHeal location is reached in the "RunAwayAndHeal" sequence in BT. The Healing is handled through the AttributeComponent and ApplyHealthChange().
Added AttributeComponent, which handles the attributes like health. Within this Class, we declared a new multicast delegate FOnHealthChanged which broadcasts the Instigator actor, Health, and Delta.
Added an SGameModeBase which handles spawning of the AI, using the Query_FindBotSpawn. The query chooses the Best Location 1000 meters away, which had a broken line of sight from the QueryContext_AllPlayers, i.e., All players. The Query is initiated on start play by a looping timer that calls the SpawnBotTimeElapsed, which in turn runs the UEnvQueryManger::RunEQSQuery() where we specify the query to run, which we exposed to be set in the blueprint child of the game mode by declaring a UEnvQuery* SpawnBotQuery. This function runs asynchronously, but immediately returns a handle UEnvQueryInstanceBlueprintWrapper* which we store in a QueryInstance, and in the next line, we add a function to a delegate that is called once the RunEQSQuery() is finished, we assign OnQueryCompleted() to this delegate. The RunEQSQuery() is executed only if the number of bots does not exceed the MaxBotCount local variable. We also check if the DifficultyCurve is valid, and if it is, we set the value of the MaxBotCount to the time value of the curve. This way, depending on the game time, the number of bots allowed to be spawned changes. OnQueryCompleted() has a fitting signature and will check if QueryStatus failed, and if it did, it will early return and log a debug message. We will store QueryInstance->GetResultAsLocations() in an array of locations, and we will check if there is at least one valid member of the array, and if there is one, we will spawn an actor of the MinionClass at that location. MinionClass is another variable of the TSubclassOf<AActor> type, which we will set in a blueprint child of the GameMode. Finally, we draw the debug sphere on the location where the AI spawned.

-Fixed the bug where the MagicProjectile, when spawned by the AI, would collide with the mesh by setting the spawn parameters and setting the instigator to the AI pawn within the SBTTask_RangedAttack.

-Added Health Bar Widget for the SACharacter